Fix bill total in single licence bill page (#524)
https://eaflood.atlassian.net/browse/WATER-4132

An issue was found with how credits were displayed that we sorted; [Fix display of credit transactions in bill views](https://github.com/DEFRA/water-abstraction-system/pull/522). Or we thought we did. 😩

The bill information has a total which needs to be `£100.00 credit`. The transactions table beneath also has a total but it needs to display `-£100.00`.

To build the single licence bill view we combine the output from `BillPresenter` with `ViewBillLicencePresenter`. It just so happens that both return a `total:` property. 🤦

This means one is overriding the other leading to an issue on the single licence bill page. This change fixes it.