Fix bill total wrapping when a credit
https://eaflood.atlassian.net/browse/WATER-4131

We have found when the value of a credit gets beyond a few pounds that the current view will wrap it even though there appears to be plenty of space.

So, in this change we tweak the [GDS layout class](https://design-system.service.gov.uk/styles/layout/#common-layouts) being used to give us a little extra space. The value will still wrap if it's _really_ large. But in most cases now it should be on one line.