Rename custom SCSS class
https://eaflood.atlassian.net/browse/WATER-4131

When working on [Fix display of credit transactions in bill views](https://github.com/DEFRA/water-abstraction-system/pull/522) after adding the missing `-` to credit values we found the UI was wrapping it _after_ the sign, for example

```
      -
£126.74
```

We knew we could fix this by updating the CSS for the class to include `white-space: nowrap;`. The issue was where do we add it. We wasted a bunch of time trying to find the class `govuk-table__cell--total` in the design system code. Finally, we tracked it now to a custom class we'd added.

This change is about not making that mistake again and trying to set a precedent. From now on, if we have to add out own custom class we should ***not*** prefix it with `govuk` so we can distinguish between our own code and that provided by GDS.