Move `DetermineChargePeriodService`
https://eaflood.atlassian.net/browse/WATER-4188

Whilst working on creating the two-part-tariff billing service we have been able to re-use the `DetermineChargePeriodService` that was originally built for supplementary billing.

As this service is now going to be used for both supplementary and 2PT it's location is going to be moved up 1 level from `supplmentary` to `bill-runs` to reflect this.