Refactor existing service to create a new one
https://eaflood.atlassian.net/browse/WATER-4188

As part of the work we have been doing on two-part-tariff. We have found some code that we created for supplementary billing can be re-used for 2PT.

This PR will move the code from the `CalculateAuthorisedAndBillableDaysService` that is responsible for determining the abstraction period, and move it to a new service `DetermineAbstractionPeriodService`.

Once extracted out this function can then also be used by the 2PT process as well as supplementary.