Rename billing presenters to bill-runs (#536)
It was spotted that `app/services/billing` was inconsistent with our naming convention so we [sorted that](https://github.com/DEFRA/water-abstraction-system/pull/516).

Only we didn't spot that `services/billing` was using things in `presenters/billing`. Doh!

So, this change also corrects the presenters folder to, hopefully, get everything in line.