Add legacy migrations to add defaults
When building our views we are only pulling across data that is actually used. There is data in the `returns` and `lines` tables that is not used but is Not Nullable. We are therefore setting a default value for these fields rather than add useless data to our views which we then have to keep populating.