Rename bill presenters to match convention (#538)
https://eaflood.atlassian.net/browse/WATER-4131

We've realised that the presenters we built to support the new view bill, and bill licence pages don't follow the convention we've been following. Principally, we start with an adjective. For example, `view` or `create`. We do this to give new maintainers a sense of what context the presenter is used in.

So, this is a housekeeping change to rename the presenters and bring them in line.