Add new view bill run endpoint and template
https://eaflood.atlassian.net/browse/WATER-4223

We have been working on implementing a new version of the bill page. The current one tries to display all transaction details for all licences on a single page when the bill is selected.

In the larger annual bill runs this might equate to more than 150 licences and their transactions which means the page is failing to load.

All that work has been done in this repo as part of our ongoing efforts to migrate away from the legacy code. It is in the final throws of testing and will soon be released.

To get to the new pages you have to pass through the existing bill run summary page. What our work has highlighted is this page leaves a lot to desire! Again, the larger the bill run the slower it is, noticeably so. Though it at least always loads 🧌. It also is inconsistent. It has a different layout depending on the bill run type, and one of those depends on tabs which is an incorrect use of the component (and 🤢).

So, we want to remove this remaining 'bump' in our billing user journey and replace it with a 'better' one. This is the final change in a series to support that, which adds the new template and endpoint.