Rename existing bill-runs template to review.njk (#545)
https://eaflood.atlassian.net/browse/WATER-4205

In [Linking 2PT SROC review page](https://github.com/DEFRA/water-abstraction-system/pull/504) we added a stand-in view to confirm that if you click a bill run in `REVIEW` in the bill runs page of the internal UI you will get to a page we intend to build out and supporting reviewing the match & allocate results.

What we should have spotted is that the template was named `view.njk` not `review.njk`. We need `view.njk` as that would be the default template name for viewing a bill run. So, this quick change renames the template to `review.njk` to match the controller name and endpoint.