Double doh!
Another thing we missed. Because of how much we are trying to display in our tables we have to add our own custom widths that go a little smaller than the ones provided by the design system (`one-quarter` is the smallest they go to). These are needed in the bill run page when we have both tables visible and want to ensure the columns for both line up. Added to that supplementary bills feature an extra column hence the need for 3 custom widths.