Remove back-link nunjucks includes (#540)
https://eaflood.atlassian.net/browse/WATER-4131

We created a generic back-link [includes file](https://mozilla.github.io/nunjucks/templating.html#include) for our Nunjucks template as part of building the new billing pages. It was initially built when we thought the requirement was to return users back to where they came from. But working with our designers we've all agreed that back-links in the service should take users to specific places. They should _not_ just be a stand-in for the browser back button.

This means we never intend to use this includes file and we don't want those unaware of this decision to inadvertently make use of it. So, this change yeet's it from the codebase!