Add fetch bill run service for view bill run page (#537)
https://eaflood.atlassian.net/browse/WATER-4223

This service has been added to get the data needed for a new bill run page that will replace the existing one in the legacy service.

The page displays top-level details about the bill run (date created type, region etc) and then a summary for each bill linked to the bill run. The complexity is

- each bill in the table is expected to include all the licence references linked to the bill
- if a licence linked to a bill is marked as `water_undertaker` the bill should be placed in a separate table of 'Water companies' and all other bills will then go in an 'Other abstractors' table

The service needs to fetch the top-level data as well as the bill summary data.