Create `water` schema views
https://eaflood.atlassian.net/browse/WATER-4188

As part of the work we have been doing on two-part tariff we are going to be creating all our new tables in a single schema `public`.

We have also decided that when there is a legacy table that we are still going to need for SROC. We are going to create a View of that table in the `public` schema and correct any issues with naming conventions, unused fields etc in the view. The first of these views that represents the old `returns` tables have been created in https://github.com/DEFRA/water-abstraction-system/pull/531

In this PR we will build the migration scripts to create the Views we need for SROC in the `water` schema.