Move DetermineIssuesService to /check
The DetermineIssuesService was a starter-for-ten place to work out how we determine the issues on returns and elements and how we set the status for elements and licences.

However, it was done on the data we use to match and allocate when it will actually need to be based on review result data. Plus this isn't the place where it will be used. It will be when we load the review page.

Where it could be really handy right now though is in our existing check service. We can use it against the data we are returning there and confirm we are allocating issues and statuses correctly. We can then still refer to it when it comes to implementing the 'final' version of the logic.