Create `idm` schema views (#555)
https://eaflood.atlassian.net/browse/WATER-4057

As part of the work we have been doing on two-part tariff we are going to be creating all our new tables in the default `public` schema.

We have also decided that when there is a legacy table that we are still going to need we will create a [View](https://www.postgresql.org/docs/current/sql-createview.html) of it in the `public` schema. This allows us to correct any issues with naming conventions, strip out unused fields, and join entities that are currently sat in different schemas. The first example of this approach was done in PR #531 .

This change adds the migrations needed to create views for the tables we are currently using from `idm` schema.

> We'll follow this up with new models and test helpers in another change. The final step is then to refactor the existing code to use the new models and delete the old legacy ones.