Fix billing acc. address error when town is empty (#558)
Fix billing acc. address error when town is empty
    
https://eaflood.atlassian.net/browse/WATER-4231

Spotted whilst trying to change the billing account address locally. We were adding an FAO to a billing account address that turns out didn't have the `town` field populated. When we tried to confirm the change we kept getting an error from the `ChargingModuleCreateCustomerChangePresenter`.

After looking into it `town` was a property we were passing to our `_truncate()` method without first checking if the value existed. We also didn't have a unit test covering this scenario.

This change fixes the presenter and adds the missing unit test.