Refactor to use new view based IDM models
https://eaflood.atlassian.net/browse/WATER-4057

As part of the work we have been doing on two-part tariff we are going to be creating all our new tables in the default `public` schema.

We have also decided that when there is a legacy table that we are still going to need we will create a [View](https://www.postgresql.org/docs/current/sql-createview.html) of it in the `public` schema. This allows us to correct any issues with naming conventions, strip out unused fields, and join entities that are currently sat in different schemas. The first example of this approach was done in PR #531 .

In [Create idm schema views](https://github.com/DEFRA/water-abstraction-system/pull/555) we created the new views. Then in [Create idm schema models and helpers](https://github.com/DEFRA/water-abstraction-system/pull/563) we added the new models and helpers that use them.

This is the final step in the process, refactoring any use of the legacy models to use the new ones, and deleted all the legacy based code.