Refactor to use new view based RETURNS models (#568)
https://eaflood.atlassian.net/browse/WATER-4057

As part of the work we have been doing on two-part tariff, we will be creating all our new tables in the default `public` schema.

We have also decided that when there is a legacy table that we are still going to need we will create a [View](https://www.postgresql.org/docs/current/sql-createview.html) of it in the `public` schema. This allows us to correct any issues with naming conventions, strip out unused fields, and join entities currently sat in different schemas. The first example of this approach was done in PR #531  in which views were created for the RETURNS schema.

We then created the new models and helpers that use them in [Build models for returns based on new Views](https://github.com/DEFRA/water-abstraction-system/pull/533).

This is the final step in the process, we are refactoring any use of the legacy models to use the new ones and deleting all the legacy-based code.