Fetch charge versions service (#546)
* https://eaflood.atlassian.net/browse/WATER-4188

This change is part of the work we have been doing on two-part-tariff. Introducing the FetchChargeVersionsService. This service will take a region code and a billing period and will return all the charge versions with there associated licences, charge references and elements. This is based on the charge version dates crossing over the billing period dates. The charge version being a status of current, the charge versions having an sroc scheme and it being in the region that was given to the service.