Add no returns required page (return requirements)
https://eaflood.atlassian.net/browse/WATER-4248

We are starting work on adding support for setting up return requirements within the service. A return requirement sets out how a licensee is expected to submit their returns, for example, daily, weekly or monthly, by what date, and using what unit of measure.

This is the very first page in the return requirements setup journey when 'no returns required' is selected.

This is just a stub page and and controls and validation will come later.