Update 2PT check endpoint with alternate changes
https://eaflood.atlassian.net/browse/WATER-4057

Way, way back we built our very first iteration of a two-part tariff 'check' endpoint. The intent was to have some way for our QA team to being verifying how we were matching and allocating returns was correct. Along the journey we realised we needed to make major shift in how things were working which has since become known as the [alternate engine](https://github.com/DEFRA/water-abstraction-system/pull/458).

This started out as a 'spike' (does _this_ look right?), then was deployed on the branch to environments for double checking, was tweaked, tweaked a bit more, tweaked again, new requests catered for etc. _Now_ we're trying to figure out how we migrate from our hack to 'production-ready' unit tested code, still in the same branch.

Recently, we lost the original alternate engine changes we made to the 'check' endpoint on that branch when we merged in `main`. That became the trigger to move out of the [alternate-2pt-engine](https://github.com/DEFRA/water-abstraction-system/tree/alternate-2pt-engine) branch those changes which are just related to the check endpoint and what the QA team need to keep an eye on things.

This should mean we don't lose those changes again. Also, the team focused on refactoring can differentiate their changes from stuff we've been doing to keep the QA's happy! 😁

So, this change updates the 'check' endpoint to be the version we've actually been using and testing against. This includes what was needed to get it to work with our recent model changes.

> When two-part tariff is completed, hopefully in the near future, we will delete all of this! 😈