Record return line to element matches
The real service won't do this because we don't care. But the verification service does, if only so we can match the scenario format. If we don't record which line matched to which return at the time of processing there is no way to be certain after the fact.

So, for now we hack our stand-in service. But when the real one gets merged in and we start using it we'll just have to accept you won't be able to see this information in the verification output any more.