Create `fetch-licences` service (#577)
https://eaflood.atlassian.net/browse/WATER-4188

As part of the work we have been doing, we've been hacking away at a branch in a mob to build the two-part-tariff service. As a result of this, we have some refactoring to do to tidy up the work we have done in our mob. We are keeping track of the refactoring to do in this issue https://github.com/DEFRA/water-abstraction-team/issues/105.

This PR will extract the `fetch-licences` service out of our hacky branch https://github.com/DEFRA/water-abstraction-system/pull/458 and into its own PR along with the unit tests.