Include volume in FetchChargeVersionsService
https://eaflood.atlassian.net/browse/WATER-4057

This is related to our work building a new SROC two-part tariff matching and allocating engine.

We've been asked to make a change to the engine. Now when it allocates returns to charge elements it should take into account the authorised volume of the charge reference the charge element is under.

For example, if we had the following scenario

- `CR01 - authorised 100`
  - `CE01 - authorised 75`
  - `CE02 - authorised 75`
- `RT01 - submitted 150`

The original understanding was that we would be able to allocate all 150ML across the 2 charge elements. Now, when the total allocation across all charge elements within a charge reference reach the charge references authorised we have to stop. So, the result would be

- `CR01 - authorised 100`
  - `CE01 - authorised 75 / allocated 75`
  - `CE02 - authorised 75 / allocated 25`
- `RT01 - submitted 150 / allocated 100`

In order to be able to do this we need to know what the authorised volume of the charge reference is. So, this change adds that to what we get back when calling `FetchChargeVersionsService`.