Add new page for add a note
https://eaflood.atlassian.net/browse/WATER-4260

We are starting work on adding support for setting up return requirements within the service. A return requirement sets out how a licensee is expected to submit their returns, for example, daily, weekly or monthly, by what date, and using what unit of measure.

This is a page for adding a note to a return requirement.

This is just a stub page and controls and validation will come later.