Boy scout - compare directly in expect
We think it's a easier to follow the test and what exactly is being compared if the expectation is in the `expect()` directly. If we needed to check that a result was repeatedly a specific value then it makes sense to put that in a variable. But in this case we are just doing it the once.

We also remove the interstitial variables which were being used to check values against. We assume this was to be able to check the ID's matched to confirm the right records were being linked. As we create the records we can set those without the need to hold things in variables.