Boy scout - delete something already tested
The main test at the top of the file already confirms this is the case. Assuming we were testing this for 'documentation' purposes it was already covered implicitly. If the current test describes, 'and the start date' + 'is after the billing period end`, were combined with 'returns no records' it is implicit that if the start date is _before_ the billing period end then we would get results.