Boy scout - reorganise test blocks
This was referred to in previous changes. We can document behaviour implicitly by explicitly testing certain cases.

For example, if we explicitly test that if the region code is different we get no results, it is implicit that if the region code is the same then we do get results. Both cases are documented, and our generic 'if all things are good we get results' provides the cover to ensure this.