Exclude ended licences in two-part tariff matching (#595)
https://eaflood.atlassian.net/browse/WATER-4057

When a two-part tariff bill run is triggered one of the first steps is fetching all the charge versions and their relevant licences that are applicable (same region and cross over the billing period).

A scenario we encountered when testing was a licence that had a 'current' charge version with a start date of 2022-04-01 but the licence was marked as lapsed in 2009.

When processed by our engine it resulted in a blank charge period, which then caused a whole bunch more errors. Initially, we made changes to cater for this scenario but having spoken with our SMEs they have agreed that ended licences can be excluded completely.

So, this change updates `FetchChargeVersionsService` to exclude ended (lapsed, expired or revoked) licences.