Fix check two-part tariff region endpoint
https://eaflood.atlassian.net/browse/WATER-4057

When we were updating the existing check endpoint to include changes made to support the alternate version of the matching and allocating engine (see [Update 2PT check endpoint with alternate changes](https://github.com/DEFRA/water-abstraction-system/pull/582)) we confirmed our working by constantly pinging the `/check/two-part-licence/{licenceId}` endpoint.

We overlooked checking `/check/two-part/{naldRegionId}` (in our defence this is a hack endpoint!) If we had we would have found that it is broken.

This change fixes it.