Add Contact info link to header
https://eaflood.atlassian.net/browse/WATER-4131

In the legacy UI there is a 'Contact information' link in the header. This takes you to a screen that where you enter a name, job title, email address, phone number and address. There is just an **Update** button and a note saying

> This will be visible to your public contacts

When we looked at the legacy code this is sourced from the `/external` side of the [water-abstraction-ui](https://github.com/DEFRA/water-abstraction-ui) code. If you log in as an external user you see it. We assumed it was therefore another bug and shouldn't be there.

But we've since learnt it is actually part of the notifications flow. When a user generates, for example, a renewal notification it will pull through their details and include them on the invite. This screen is needed to populated those details.

So, it _is_ a tacked on thing (we'd argue it should be part of the flow not some rando link in the header) but it is required. This change adds the link back in.

The only complication is the legacy code only displays it for users with a `hof_notifications` or `renewal_notifications` scope. Again, it really doesn't matter if all users see this and update there details, even if no one will see them. But we also really want to get our new bill run pages out in order to give users access to the larger bill runs. So, we replicate that feature here as well.