Amend bill count in bill run view for zero value (#602)
https://eaflood.atlassian.net/browse/WATER-4223

In testing of the new bill run view we've built a scenario was found where the count of bills we show didn't match with the number of bills on the page.

For example, the bill run had 6 credit notes and 4 invoices so we displayed "10 Supplementary bills" as the caption. But when you counted the bills listed there were 11.

We double-checked what the legacy UI was doing with this bill run and found it also displayed 6 credits and 4 invoices. But its caption was "11 supplementary bills".

Neither looks right!

After digging in we found the extra bill is a 'zero value' bill (where the credit and debit lines have value but they cancel each other out). These don't get sent to SSCL so are not included in the credit note and invoice count.

After consultation with users, they don't want these to be included in the credit note or invoice count as this will mess up their reconciliation process. But they would like the total count to be correct. The agreed solution is that when a bill run contains zero-value bills we should extend the caption, for example, "10 Supplementary bills and 1 zero value bill".

This change implements that solution.