Add README to models and migrations folders
Having just completed a refactoring of our models it is no longer as clear as it used to be where they are source from. We're also now starting to mix up models based on tables (ours) with those based on views (legacy).

So, we thought it a prudent idea to add a bit of documentation to provide a bit of context when it comes to the models and the migrations.

This change adds a README for each folder that will hopefully be useful for a new starter (and a reminder for those who have been around too long!)