Move jobs to an explicit /jobs endpoint
https://eaflood.atlassian.net/browse/WATER-4270

This is indirectly related to WATER-4270, which is about getting a better understanding of what the 'jobs' in [water-abstraction-import](https://github.com/DEFRA/water-abstraction-import) are doing.

One of the blockers to that has been simply finding them all in the code base! It's not always easy to find them, especially when the word 'job' in the legacy code base has to means and uses and are often the same thing! 😱😂

To try and avoid that happening in our code base this change moves our existing 'jobs' to a `/jobs` endpoint and tidies up the underlying structure to match. We also add a bit of info for the uninitiated to define what we mean by 'job' and explain the approach we have gone for.

We've briefly spoken about doing this as a team awhile back. We also plan to arrange for the latest job to be added ([Remove licences with charge elements with approaching time limits from billing](https://github.com/DEFRA/water-abstraction-system/pull/443)) to be setup in `production` as part of the next release. So, whilst we're reminded of this housekeeping task now seems a good time to do it.