Move jobs to an explicit /jobs endpoint (#607)
https://eaflood.atlassian.net/browse/WATER-4270

This is indirectly related to WATER-4270, which is about getting a better understanding of what the 'jobs' in [water-abstraction-import](https://github.com/DEFRA/water-abstraction-import) are doing.

One of the blockers to that has been simply finding them all in the code base! It's not always easy to find them, especially when the word 'job' in the legacy code base has two different uses and is often used for the same thing! 😱😂

To try and avoid that happening in our code base this change moves our existing 'jobs' to a `/jobs` endpoint and tidies up the underlying structure to match.

We've briefly spoken about doing this as a team a while back. We also plan to arrange for the latest job to be added ([Remove licences with charge elements with approaching time limits from billing](https://github.com/DEFRA/water-abstraction-system/pull/443)) to be set up in `production` as part of the next release. So, whilst we're reminded of this housekeeping task now seems a good time to do it.